chore (refactoring): move clean_charbuf to util, linting.
