+++ custom_components/bermuda/device_tracker.py
@@ -32,7 +32,9 @@ async def async_setup_entry(
     created_devices = []  # list of devices we've already created entities for
 
     @callback
-    def device_new(address: str, scanners: list[str]) -> None:
+    def device_new(
+        address: str, scanners: list[str]
+    ) -> None:  # pylint: disable=unused-argument
         """Create entities for newly-found device
 
         Called from the data co-ordinator when it finds a new device that needs
